feat: Rename tokeniser.py to train_tokeniser.py
