+++ src/__init__.py
@@ -1,2 +1,2 @@
 from .tokeniser_config import TokeniserConfig  # noqa
-from .tokenisers import train_tokeniser  # noqa
+from .train_tokeniser import train_tokeniser  # noqa
